Navigation ready Co_Authored_By Hector Ruvalcaba <hectormanuel_ruvalcaba@ucol.mx>
