+++ AlertCovidApp/src/index.js
@@ -1,13 +1,15 @@
 import React from 'react';
-import {View, Text} from 'react-native';   
+import {View, Text} from 'react-native';
+import {NavigationContainer} from '@react-navigation/native';  
 import LogIn from './screens/LogIn';
 import Welcome from './screens/WelcomeScreen';
+import AppNavigator from './navigations/AppNavigation'
 
 const index = () => {
   return ( 
-    <View>
-      <LogIn></LogIn>
-    </View>
+    <NavigationContainer>
+      <AppNavigator></AppNavigator>
+    </NavigationContainer>
   );
 };
